Feat: Allow MacOS menubar icon height to be overridden
Implemented via functional options, so as to be
backwards-compatible for existing importers. Allow
the height of the menubar icon on MacOS to be set,
scaling the input image proportionally to avoid
any kind of squish effect.

The default size of 16x16 is fine, but it can be
nice to use up a little more space with a 22-height
icon.