Prevent channel crash after item is removed
When an item is removed its click channel is closed. But if the item reference is still kept externally then any changes to it would add back the item, with a closed channel. Clicking on it would result in a crash.
Changes:
New items is added into menuItems. When an item is changed, before updating it is checked if the item exists in menuItems. If is not, then returns an error.